rename: 폴더명 컨벤션 맞지 않는 것 수정
