+++ src/components/workbookDetailPage/WorkbookDetailPageLayout.tsx
